+++ hmailserver/source/Server/SMTP/SMTPConnection.cpp
@@ -474,7 +474,7 @@ namespace HM
 
       // Parse the extensions 
       String sAuthParam;
-      int iEstimatedMessageSize = 0;
+      size_t iEstimatedMessageSize = 0;
       while (iterParam != vecParams.end())
       {
          String sParam = (*iterParam);
@@ -869,7 +869,7 @@ namespace HM
       transmission_buffer_->Append(pBuf->GetBuffer(), pBuf->GetSize());
 
       // We need current message size in KB
-      int iBufSizeKB = transmission_buffer_->GetSize() / 1024;
+      size_t iBufSizeKB = transmission_buffer_->GetSize() / 1024;
 
       // Clear the old buffer
       pBuf->Empty();
@@ -889,7 +889,7 @@ namespace HM
       {
 
          String sLogData;
-         int iMaxSizeDrop = IniFileSettings::Instance()->GetSMTPDMaxSizeDrop();
+         size_t iMaxSizeDrop = IniFileSettings::Instance()->GetSMTPDMaxSizeDrop();
          if (iMaxSizeDrop > 0 && iBufSizeKB >= iMaxSizeDrop) 
          {
             sLogData.Format(_T("Size: %d KB, Max size: %d KB - DROP!!"), 
@@ -1308,12 +1308,12 @@ namespace HM
       {
          // Check that buffer contains correct line endings.
          const char *pChar = pBuffer->GetCharBuffer();
-         int iBufferSize = pBuffer->GetSize();
+         size_t iBufferSize = pBuffer->GetSize();
 
          // Check from pos 3 to size-3. Not 100% sure, but
          // we don't have to worry about buffer start/endings.
 
-         for (int i = 3; i < iBufferSize - 3; i++)
+         for (size_t i = 3; i < iBufferSize - 3; i++)
          {
             const char *pCurrentChar = pChar + i;
